wip: read-only test setup
For Simon. This is intended to allow creating a file via newTable,
then transfer it to ManagedRepository via python magic, and then
use RawAccessRequest to mark it read only. This would provide a
proper integration test framework for testing the HDF5 ro issue.
