+++ components/tools/OmeroPy/test/integration/test_repository.py
@@ -9,252 +9,16 @@
 
 """
 
-import platform
-import locale
 import pytest
-from omero.testlib import ITest
+
 import omero
 
+from library import AbstractRepoTest
 from omero import CmdError
-from omero.model import NamedValue as NV
 from omero.callbacks import CmdCallbackI
 from omero.gateway import BlitzGateway
-from omero.rtypes import rbool
-from omero.rtypes import rstring
 from omero.rtypes import unwrap
 from omero.util.temp_files import create_path
-from omero_version import omero_version
-
-
-class AbstractRepoTest(ITest):
-
-    def setup_method(self, method):
-        self.unique_dir = self.test_dir()
-
-    def test_dir(self, client=None):
-        if client is None:
-            client = self.client
-        mrepo = self.getManagedRepo(client=client)
-        user_dir = self.user_dir(client=client)
-        unique_dir = user_dir + "/" + self.uuid()  # ok
-        mrepo.makeDir(unique_dir, True)
-        return unique_dir
-
-    def user_dir(self, client=None):
-        if client is None:
-            client = self.client
-        ec = client.sf.getAdminService().getEventContext()
-        return "%s_%s" % (ec.userName, ec.userId)
-
-    def all(self, client):
-        ctx = dict(client.getImplicitContext().getContext())
-        ctx["omero.group"] = "-1"
-        return ctx
-
-    def getManagedRepo(self, client=None):
-        if client is None:
-            client = self.client
-        mrepo = client.getManagedRepository()
-        assert mrepo
-        return mrepo
-
-    def createFile(self, mrepo1, filename):
-        rfs = mrepo1.file(filename, "rw")
-        try:
-            rfs.write("hi", 0, 2)
-            ofile = rfs.save()
-            return ofile
-        finally:
-            rfs.close()
-
-    def assertWrite(self, mrepo2, filename, ofile):
-        def _write(rfs):
-            try:
-                rfs.write("bye", 0, 3)
-                assert "bye" == rfs.read(0, 3)
-                # Resetting for other expectations
-                rfs.truncate(2)
-                rfs.write("hi", 0, 2)
-                assert "hi" == rfs.read(0, 2)
-            finally:
-                rfs.close()
-
-        # TODO: fileById is always "r"
-        # rfs = mrepo2.fileById(ofile.id.val)
-        # _write(rfs)
-
-        rfs = mrepo2.file(filename, "rw")
-        _write(rfs)
-
-    def assertNoWrite(self, mrepo2, filename, ofile):
-        def _nowrite(rfs):
-            try:
-                pytest.raises(omero.SecurityViolation,
-                              rfs.write, "bye", 0, 3)
-                assert "hi" == rfs.read(0, 2)
-            finally:
-                rfs.close()
-
-        rfs = mrepo2.fileById(ofile.id.val)
-        _nowrite(rfs)
-
-        rfs = mrepo2.file(filename, "r")
-        _nowrite(rfs)
-
-        # Can't even acquire a writeable-rfs.
-        pytest.raises(omero.SecurityViolation,
-                      mrepo2.file, filename, "rw")
-
-    def assertDirWrite(self, mrepo2, dirname):
-        self.createFile(mrepo2, dirname + "/file2.txt")
-
-    def assertNoDirWrite(self, mrepo2, dirname):
-        # Also check that it's not possible to write
-        # in someone else's directory.
-        pytest.raises(omero.SecurityViolation,
-                      self.createFile, mrepo2, dirname + "/file2.txt")
-
-    def assertNoRead(self, mrepo2, filename, ofile):
-        pytest.raises(omero.SecurityViolation,
-                      mrepo2.fileById, ofile.id.val)
-        pytest.raises(omero.SecurityViolation,
-                      mrepo2.file, filename, "r")
-
-    def assertRead(self, mrepo2, filename, ofile, ctx=None):
-        def _read(rfs):
-            try:
-                assert "hi" == rfs.read(0, 2)
-            finally:
-                rfs.close()
-
-        rfs = mrepo2.fileById(ofile.id.val, ctx)
-        _read(rfs)
-
-        rfs = mrepo2.file(filename, "r", ctx)
-        _read(rfs)
-
-    def assertListings(self, mrepo1, unique_dir):
-        assert [unique_dir + "/b"] == mrepo1.list(unique_dir + "/")
-        assert [unique_dir + "/b/c"] == mrepo1.list(unique_dir + "/b/")
-        assert [
-            unique_dir + "/b/c/file.txt"] == mrepo1.list(unique_dir + "/b/c/")
-
-    def raw(self, command, args, client=None):
-        if client is None:
-            client = self.client
-        mrepo = self.getManagedRepo(self.client)
-        obj = mrepo.root()
-        sha = obj.hash.val
-        raw_access = omero.grid.RawAccessRequest()
-        raw_access.repoUuid = sha
-        raw_access.command = command
-        raw_access.args = args
-        handle = client.sf.submit(raw_access)
-        return CmdCallbackI(client, handle)
-
-    def assertPasses(self, cb, loops=10, wait=500):
-        cb.loop(loops, wait)
-        rsp = cb.getResponse()
-        if isinstance(rsp, omero.cmd.ERR):
-            raise Exception(rsp)
-        return rsp
-
-    def assertError(self, cb, loops=10, wait=500):
-        cb.loop(loops, wait)
-        rsp = cb.getResponse()
-        assert isinstance(rsp, omero.cmd.ERR)
-
-    def create_test_dir(self):
-        folder = create_path(folder=True)
-        (folder / "a.fake").touch()
-        (folder / "b.fake").touch()
-        return folder
-
-    def create_fileset(self, folder):
-        fileset = omero.model.FilesetI()
-        for f in folder.files():
-            entry = omero.model.FilesetEntryI()
-            entry.setClientPath(rstring(str(f.abspath())))
-            fileset.addFilesetEntry(entry)
-
-        # Fill version info
-        system, node, release, version, machine, processor = platform.uname()
-
-        clientVersionInfo = [
-            NV('omero.version', omero_version),
-            NV('os.name', system),
-            NV('os.version', release),
-            NV('os.architecture', machine)
-        ]
-        try:
-            clientVersionInfo.append(
-                NV('locale', locale.getdefaultlocale()[0]))
-        except:
-            pass
-
-        upload = omero.model.UploadJobI()
-        upload.setVersionInfo(clientVersionInfo)
-        fileset.linkJob(upload)
-        return fileset
-
-    def create_settings(self):
-        settings = omero.grid.ImportSettings()
-        settings.doThumbnails = rbool(True)
-        settings.noStatsInfo = rbool(False)
-        settings.userSpecifiedTarget = None
-        settings.userSpecifiedName = None
-        settings.userSpecifiedDescription = None
-        settings.userSpecifiedAnnotationList = None
-        settings.userSpecifiedPixels = None
-        settings.checksumAlgorithm = omero.model.ChecksumAlgorithmI()
-        settings.checksumAlgorithm.value = omero.rtypes.rstring("SHA1-160")
-        return settings
-
-    def upload_folder(self, proc, folder):
-        ret_val = []
-        for i, fobj in enumerate(folder.files()):  # Assuming same order
-            rfs = proc.getUploader(i)
-            try:
-                f = fobj.open()
-                try:
-                    offset = 0
-                    block = []
-                    rfs.write(block, offset, len(block))  # Touch
-                    while True:
-                        block = f.read(1000 * 1000)
-                        if not block:
-                            break
-                        rfs.write(block, offset, len(block))
-                        offset += len(block)
-                    ret_val.append(self.client.sha1(fobj.abspath()))
-                finally:
-                    f.close()
-            finally:
-                rfs.close()
-        return ret_val
-
-    def fullImport(self, client):
-        """
-        Re-usable method for a basic import
-        """
-        mrepo = self.getManagedRepo(client)
-        folder = self.create_test_dir()
-        fileset = self.create_fileset(folder)
-        settings = self.create_settings()
-
-        proc = mrepo.importFileset(fileset, settings)
-        try:
-            return self.assertImport(client, proc, folder)
-        finally:
-            proc.close()
-
-    def assertImport(self, client, proc, folder):
-        hashes = self.upload_folder(proc, folder)
-        handle = proc.verifyUpload(hashes)
-        cb = CmdCallbackI(client, handle)
-        rsp = self.assertPasses(cb)
-        assert 1 == len(rsp.pixels)
-        return rsp
 
 
 class TestRepository(AbstractRepoTest):
